requirements.txt
appnope==0.1.0
blinker==1.4
chardet==2.3.0
click==6.6
decorator==4.0.9
dominate==2.2.0
Flask==0.10.1
Flask-Bootstrap==3.3.5.6
Flask-DebugToolbar==0.10.0
flask-mongoengine==0.7.1
flask-paginate==0.4.1
Flask-RQ==0.2
Flask-Runner==2.1.1
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.12
Flask-Mail==0.9.0
futures==3.0.3
google-api-python-client==1.4.2
gunicorn==19.4.5
html5lib==0.9999999
httplib2==0.9.2
ipython==4.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
json2html==1.0.1
line-profiler==1.0
lxml==3.5.0
MarkupSafe==0.23
mongoengine==0.10.6
nose==1.3.7
oauth2client==1.5.2
ordereddict==1.1
path.py==8.1.2
PbPython==0.1.9
pexpect==4.0.1
pickleshare==0.5
polib==1.0.7
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pymongo==3.2.2
pysrt==1.0.1
pyter==0.2.2.1
python-dateutil==2.4.2
python-dotenv==0.3.0
python-magic==0.4.10
redis==2.10.5
requests[security]==2.9.1
rq==0.5.6
rsa==3.4.2
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.1
uritemplate==0.6
Werkzeug==0.11.8
WTForms==2.1
mafan==0.2.10
sklearn==0.0
scipy==1.1.0
